Changed to reflect new guidelines.

+++ TAO/examples/Simple/time/run_test.pl
@@ -9,15 +9,25 @@ unshift @INC, '../../../../bin';
 require ACEutils;
 require Process;
 
+$status = 0;
 $iorfile = "time.ior";
 
 $SV = Process::Create ($EXEPREFIX."server$EXE_EXT", "-o $iorfile ");
 
-ACE::waitforfile ($iorfile);
+if (ACE::waitforfile_timed ($iorfile, 5) == -1) {
+  print STDERR "ERROR: timedout waiting for file <$iorfile>\n";
+  $SV->Kill (); $SV->TimedWait (1);
+  exit 1;
+}
 
-$status  = system ($EXEPREFIX."client$EXE_EXT -f $iorfile -x");
+$client = Process::Create ($EXEPREFIX."client$EXE_EXT", "-f $iorfile -x");
+if ($client->TimedWait (60) == -1) {
+  print STDERR "ERROR: client timedout\n";
+  $status = 1;
+  $client->Kill (); $client->TimeWait (1);
+}
 
-$SV->Kill (); $SV->Wait ();
+$SV->Kill (); $SV->TimedWait (1);
 
 unlink $iorfile;
